+++ index.d.ts
@@ -220,6 +220,18 @@ export interface PurchaserInfo {
      * in Android
      */
     readonly originalApplicationVersion: string | null;
+    /**
+     * Returns the purchase date for the version of the application when the user bought the app.
+     * Use this for grandfathering users when migrating to subscriptions.
+     */
+    readonly originalPurchaseDate: string | null;
+    /**
+     * URL to manage the active subscription of the user. If this user has an active iOS
+     * subscription, this will point to the App Store, if the user has an active Play Store subscription
+     * it will point there. If there are no active subscriptions it will be null.
+     * If there are multiple for different platforms, it will point to the device store.
+     */
+    readonly managementURL: string | null;
 }
 export interface PurchasesProduct {
     /**
@@ -523,12 +535,14 @@ export default class Purchases {
     static INTRO_ELIGIBILITY_STATUS: typeof INTRO_ELIGIBILITY_STATUS;
     /**
      * Sets up Purchases with your API key and an app user id.
-     * @param {string} apiKey RevenueCat API Key. Needs to be a String
+     * @param {String} apiKey RevenueCat API Key. Needs to be a String
      * @param {String?} appUserID An optional unique id for identifying the user. Needs to be a string.
      * @param {Boolean?} observerMode An optional boolean. Set this to TRUE if you have your own IAP implementation and want to use only RevenueCat's backend. Default is FALSE.
+     * @param {String?} userDefaultsSuiteName An optional string. iOS only. Set this to use a specific NSUserDefaults suite for RevenueCat.
+     * This might be handy if you are deleting all NSUserDefaults in your app and leaving RevenueCat in a bad state.
      * @returns {Promise<void>} Returns when setup completes
      */
-    static setup(apiKey: string, appUserID?: string | null, observerMode?: boolean): any;
+    static setup(apiKey: string, appUserID?: string | null, observerMode?: boolean, userDefaultsSuiteName?: string): any;
     /**
      * @param {Boolean} allowSharing Set this to true if you are passing in an appUserID but it is anonymous, this is true by default if you didn't pass an appUserID
      * If an user tries to purchase a product that is active on the current app store account, we will treat it as a restore and alias
@@ -724,13 +738,13 @@ export default class Purchases {
     static getPaymentDiscount(product: PurchasesProduct, discount: PurchasesDiscount): Promise<PurchasesPaymentDiscount | undefined>;
     /**
      * Invalidates the cache for purchaser information.
-     * 
+     *
      * Most apps will not need to use this method; invalidating the cache can leave your app in an invalid state.
      * Refer to https://docs.revenuecat.com/docs/purchaserinfo#section-get-user-information for more information on
      * using the cache properly.
-     * 
-     * This is useful for cases where purchaser information might have been updated outside of the
-     * app, like if a promotional subscription is granted through the RevenueCat dashboard.
+     *
+     * This is useful for cases where purchaser information might have been updated outside of the app, like if a
+     * promotional subscription is granted through the RevenueCat dashboard.
      */
     static invalidatePurchaserInfoCache(): void;
     /**
@@ -770,5 +784,6 @@ export default class Purchases {
      * @param pushToken null will delete the subscriber attribute.
      */
     static setPushToken(pushToken: string | null): void;
+    static setProxyURLString(url: string): void;
 }
 export {};