+++ ios/RNPurchases.m
@@ -33,15 +33,16 @@ NSString *RNPurchasesShouldPurchasePromoProductEvent = @"Purchases-ShouldPurchas
 RCT_EXPORT_MODULE();
 
 RCT_EXPORT_METHOD(setupPurchases:(NSString *)apiKey
-                  appUserID:(NSString *)appUserID
+                  appUserID:(nullable NSString *)appUserID
                   observerMode:(BOOL)observerMode
+                  userDefaultsSuiteName:(nullable NSString *)userDefaultsSuiteName
                   resolve:(RCTPromiseResolveBlock)resolve
                   reject:(RCTPromiseRejectBlock)reject)
 {
     [RCPurchases configureWithAPIKey:apiKey
                            appUserID:appUserID
                         observerMode:observerMode
-                        userDefaults:nil
+               userDefaultsSuiteName:userDefaultsSuiteName
                       platformFlavor:self.platformFlavor
                platformFlavorVersion:self.platformFlavorVersion];
     RCPurchases.sharedPurchases.delegate = self;
@@ -61,7 +62,7 @@ RCT_EXPORT_METHOD(setFinishTransactions:(BOOL)finishTransactions)
 
 RCT_EXPORT_METHOD(addAttributionData:(NSDictionary *)data
                   forNetwork:(NSInteger)network
-                  forNetworkUserId:(NSString * _Nullable)networkUserId)
+                  forNetworkUserId:(nullable NSString *)networkUserId)
 {
     [RCCommonFunctionality addAttributionData:data network:network networkUserId:networkUserId];
 }
@@ -124,7 +125,7 @@ RCT_REMAP_METHOD(getAppUserID,
     resolve([RCCommonFunctionality appUserID]);
 }
 
-RCT_EXPORT_METHOD(createAlias:(NSString * _Nullable)newAppUserID
+RCT_EXPORT_METHOD(createAlias:(nullable NSString *)newAppUserID
                   resolve:(RCTPromiseResolveBlock)resolve
                   reject:(RCTPromiseRejectBlock)reject)
 {
@@ -132,7 +133,7 @@ RCT_EXPORT_METHOD(createAlias:(NSString * _Nullable)newAppUserID
                        completionBlock:[self getResponseCompletionBlockWithResolve:resolve reject:reject]];
 }
 
-RCT_EXPORT_METHOD(identify:(NSString * _Nullable)appUserID
+RCT_EXPORT_METHOD(identify:(nullable NSString *)appUserID
                   resolve:(RCTPromiseResolveBlock)resolve
                   reject:(RCTPromiseRejectBlock)reject) {
     [RCCommonFunctionality identify:appUserID
@@ -228,6 +229,11 @@ RCT_EXPORT_METHOD(setDisplayName:(NSString *)displayName)
 RCT_EXPORT_METHOD(setPushToken:(NSString *)pushToken)
 {
     [RCCommonFunctionality setPushToken:pushToken];
+}
+
+RCT_EXPORT_METHOD(setProxyURLString:(nullable NSString *)proxyURLString)
+{
+    [RCCommonFunctionality setProxyURLString:proxyURLString];
 }